+++ table/filter_block.h
@@ -108,14 +108,13 @@ class FilterBlockReader {
 
   bool whole_key_filtering() const { return whole_key_filtering_; }
 
-  int GetLevel() const { return level_; }
-  void SetLevel(int level) { level_ = level; }
-
   // convert this object to a human readable form
   virtual std::string ToString() const {
     std::string error_msg("Unsupported filter \n");
     return error_msg;
   }
+
+  virtual void CacheDependencies(bool pin) {}
 
  protected:
   bool whole_key_filtering_;